Uploading my Final Project for Client-Side Web Programming
This is my final project from Client-Side Web Programming, one thing to note is that the "node_modules" file is missing here. this is needed to run the Node server, but since it is a few hundred files it cannot be uploaded. But if you already have Node set up then you should be good to go on your own.