Dungeons and Dragons dice roller application
This is a simple dice roller application written in Java and meant to be used when playing Dungeons and Dragons. It's a pretty simple app, and doesn't currently have a GUI to go along with it, so it's terminal based. A GUI was planned but wrote this back in CS 110 and didn't have the time or knowledge when writing this up, do plan on coming back and at least getting a GUI going!