Round Robin Bracket
This is a round robin bracket that can be used by multiple people at once. You enter the number of teams playing, and each team is given a number between zero and number of teams - 1. As each team plays a game you will click on the given index, each click being a possible outcome of the game. Try it out!