Wolf, Goat, and Cabbage Riddle
Included this purely for the riddle, it's kind of fun. So the goal is to get the wolf, the goat, and the cabbage to the other side of the river. But there's a catch, if the goat and the cabbage are left alone (without you and the boat around) the goat eats the cabbage. If the wolf is left alone with the goat then the wolf will eat the goat. But if you are around nothing will be eaten, clicking the button associated with any three items "preps" them to move then pressing the boat button will move it to the other side. Can you get them across without something being eaten?